+++ src/rust/jit_instructions.rs
@@ -92,7 +92,7 @@ pub fn jit_handle_prefix(ctx: &mut JitContext, instr_flags: &mut u32) {
 
 pub fn jit_handle_segment_prefix(segment: u32, ctx: &mut JitContext, instr_flags: &mut u32) {
     dbg_assert!(segment <= 5);
-    ctx.cpu.prefixes |= segment + 1;
+    ctx.cpu.prefixes |= segment as u8 + 1;
     jit_handle_prefix(ctx, instr_flags)
 }
